requirements.txt
amqp==2.4.2
APScheduler==3.6.0
bcrypt==3.1.6
billiard==3.5.0.5
celery==4.2.2
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudinary==1.15.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
coverage-badge==1.0.1
dj-database-url==0.5.0
Django==2.1.7
django-apscheduler==0.2.13
django-cors-headers==3.1.0
django-filter==2.2.0
django-heroku==0.3.1
django-redis==4.10.0
django-stubs==1.1.0
djangorestframework==3.9.2
djangorestframework-camel-case==1.1.1
djangorestframework-jwt==1.11.0
djangorestframework-stubs==0.3.0
drf-yasg==1.15.0
Flask==1.0.2
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.1
kombu==4.3.0
locustio==0.11.0
Markdown==3.1
MarkupSafe==1.1.1
mock==2.0.0
model-mommy==1.6.0
msgpack==0.6.1
mypy==0.720
mypy-extensions==0.4.1
pbr==5.1.3
psycopg2==2.7.7
pycparser==2.19
Pygments==2.3.1
PyJWT==1.7.1
python-coveralls==2.9.2
python-dotenv==0.10.3
pytz==2018.9
PyYAML==5.1.1
pyzmq==18.0.1
redis==3.2.1
requests==2.21.0
ruamel.yaml==0.15.91
ruamel.yaml.clib==0.1.2
simplejson==3.16.0
six==1.12.0
typed-ast==1.4.0
typing-extensions==3.7.4
tzlocal==1.5.1
uritemplate==3.0.0
urllib3==1.24.1
vine==1.3.0
Werkzeug==0.15.2
whitenoise==4.1.2
